Adjustments to conformance report
Changed the "Result" column to "Conformant" with values "Yes", "No", and
"Unknown". The latter reflects the case where an unexpected error occurs
preventing conformance from being determined. The column value can be
augmented with additional text describing the reason, eg for "Unknown"
the error details would be included.

The test specs report non-conformance via a description function that
is passed to Gomaga assertions that is lazily evaluated by the assertion
only if a failure occurs. The function adds a report entry named
"non-conformant" with an optional message string that is used by the
reporting backend to distinguish between a spec failure due to
non-conformance or one due to an unexpected error/failure.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>