Add conformance tests for EndpointSlices
Look for an MCS EndpointSlice using the criteria whereby the
"endpointslice.kubernetes.io/managed-by" label is not equal to
endpointslice-controller.k8s.io" or any MCS label is present.
If not found then fail the test as non-conformant. Otherwise
verify the required MCS labels are correctly set.

The tests are labeled with "EndpointSlice" so, if EndpointSlices
aren't supported by the implmentation, the user can use the
"EndpointSlice" label to filter out the tests.

Fixes https://github.com/kubernetes-sigs/mcs-api/issues/70

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>