Fix Go linting
The switch to the tools module for tool dependency tracking broke
verify-golint.sh (golint was run from tools by "go -C tools" but the
packages were no longer accessible).

This fixes the above by converting the packages to relative
references.

It also ensures that output sent to stderr is not ignored, so that
future errors of this type don't go undetected. It also ignores the
golint error about OptionalLabel; documenting that isn't very useful.

Signed-off-by: Stephen Kitt <skitt@redhat.com>