apis/v1alpha1: add target port to ServiceImport ports
This is adding target port to v1alpha1 API that was missing in the ports
definitions. In implementations that relies on derived services this
would allow to correctly propagate that information to its derived
Service.

This is also in sync with the current planned change to v1alpha2 which
uses directly the core Service ports in its struct definition which
already includes target ports.

Signed-off-by: Arthur Outhenin-Chalandre <arthur.outhenin-chalandre@ledger.fr>