Conformance tests: improve awaitServiceImport
We can improve it by utilizing the form of 'Eventually' that
accepts a function that that takes a single Gomega argument
which is used to make assertions. 'Eventually' succeeds only if all
the assertions in the polled function pass. This is simpler than
first polling to find the ServiceImport based on initial checks
then making separate assertions to provide better error output
that typically overlap with the initial checks.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>