Adjust the service type conflict conformance test
...to only verify a Conflict condition is applied on each cluster.

In Submariner, we use the type of the first observed ServiceExport
for the derived multi-cluster service. So if the service is exported
in another cluster and its actual timestamp happens to be older than
the first, we don't change the derived service to avoid disruption.
This is in accordance with the KEP which recently augmented the conflict
resolution policy to allow for favoring maintaining service continuity
and avoiding potentially disruptive changes.

So given that, I relaxed the test so it doesn't verify the service type.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>