Ensure first cluster exports service before second cluster
...in two cluster conformance tests. Previously, the tests exported
the first cluster in JustBeforeEach but that was executed after the
JustBeforeEach in the twoClusterTestDriver so the second cluster was
actually exported first.

In addition, the verification of service type was put back in for the
service type conflict conformance test. This had been removed by
https://github.com/kubernetes-sigs/mcs-api/pull/119. To make the
conflict resolution deterministic for an implementation that uses
the first observed ServiceExport, the test setup now ensures the
ServiceImport is created before deploying on the second cluster.

Signed-off-by: Tom Pantelis <tompantelis@gmail.com>