conformance: skip empty endpoints in awaitMCSEndpointSlice
Kubernetes can create empty EndpointSlice if a service select no pods
which means MCS implementation should be able to create similar
EndpointSlice and the conformances tests should be able to not fail if
such EndpointSlice exists.

This is affecting the Cilium implementation that create some empty
EndpointSlice on the StatefulSet & DNS test which makes the conformance
test fails.

This commit is thus simply skipping any empty EndpointSlice.

Signed-off-by: Arthur Outhenin-Chalandre <git@mrfreezeex.fr>