conformance: include SpecRef for skipped tests
Skipped tests doesn't include SpecRef because it's inside the Specify
clause that may be skipped at the CLI level.

This commit introduce a function SpecifyWithSpecRef that wrap the
Specify function to include the spec in a global map and then restore
the report entry within ReportBeforeEach.

It doesn't change any functional things in the tests itself but we have
to unfortunately indent everything to make up for this new arg being in
a new line.

Signed-off-by: Arthur Outhenin-Chalandre <git@mrfreezeex.fr>